fix: drop to Date across the board
